tests/integration_tests/requirements.txt
nose
nose-progressive
nose-testconfig
parameterized
requests
